move tests to dedicated file
