+++ tests/test_0692_fsspec.py
@@ -8,42 +8,6 @@ import uproot.source.fsspec
 import skhep_testdata
 
 
-def test_url_split():
-    for input_url, result in [
-        (
-            "https://github.com/scikit-hep/scikit-hep-testdata/raw/v0.4.33/src/skhep_testdata/data/uproot-issue121.root:Events",
-            (
-                "https://github.com/scikit-hep/scikit-hep-testdata/raw/v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
-                "Events",
-            ),
-        ),
-        (
-            "https://github.com/scikit-hep/scikit-hep-testdata/raw/v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
-            (
-                "https://github.com/scikit-hep/scikit-hep-testdata/raw/v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
-                None,
-            ),
-        ),
-        (
-            "github://scikit-hep:scikit-hep-testdata@v0.4.33/src/skhep_testdata/data/uproot-issue121.root:Events",
-            (
-                "github://scikit-hep:scikit-hep-testdata@v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
-                "Events",
-            ),
-        ),
-        (
-            "github://scikit-hep:scikit-hep-testdata@v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
-            (
-                "github://scikit-hep:scikit-hep-testdata@v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
-                None,
-            ),
-        ),
-    ]:
-        url, object = uproot._util.file_object_path_split(input_url)
-        assert url == result[0]
-        assert object == result[1]
-
-
 @pytest.mark.network
 def test_open_fsspec_http():
     with uproot.open(
