+++ tests/test_0976_path_object_split.py
@@ -1,0 +1,39 @@
+import pytest
+
+import uproot
+
+
+def test_url_split():
+    for input_url, result in [
+        (
+            "https://github.com/scikit-hep/scikit-hep-testdata/raw/v0.4.33/src/skhep_testdata/data/uproot-issue121.root:Events",
+            (
+                "https://github.com/scikit-hep/scikit-hep-testdata/raw/v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
+                "Events",
+            ),
+        ),
+        (
+            "https://github.com/scikit-hep/scikit-hep-testdata/raw/v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
+            (
+                "https://github.com/scikit-hep/scikit-hep-testdata/raw/v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
+                None,
+            ),
+        ),
+        (
+            "github://scikit-hep:scikit-hep-testdata@v0.4.33/src/skhep_testdata/data/uproot-issue121.root:Events",
+            (
+                "github://scikit-hep:scikit-hep-testdata@v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
+                "Events",
+            ),
+        ),
+        (
+            "github://scikit-hep:scikit-hep-testdata@v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
+            (
+                "github://scikit-hep:scikit-hep-testdata@v0.4.33/src/skhep_testdata/data/uproot-issue121.root",
+                None,
+            ),
+        ),
+    ]:
+        url, object = uproot._util.file_object_path_split(input_url)
+        assert url == result[0]
+        assert object == result[1]
